requirements.txt
chromadb==1.0.15
fastapi==0.116.1
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.72
matplotlib==3.10.5
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
seaborn==0.13.2
streamlit==1.48.0
uvicorn==0.35.0
python-multipart==0.0.6
